requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
cryptography>=3.0
defusedxml==0.6.0
Django==2.2.13
django-allauth==0.40.0
django-braces==1.14.0
django-cors-headers==3.1.1
django-oauth-toolkit==1.3.2
django-rest-auth==0.9.5
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.11.0
idna==2.10
mccabe==0.6.1
oauthlib==3.1.0
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
urllib3==1.25.9
